requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.5.25
coloredlogs==15.0.1
dataclasses-json==0.5.7
decorator==5.1.1
duckdb==0.8.0
entrypoints==0.4
exceptiongroup==1.1.1
fastapi==0.96.0
flatbuffers==23.5.26
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-aiplatform==1.25.0
google-cloud-bigquery==3.11.0
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.48.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
langchain==0.0.188
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
posthog==3.0.1
proto-plus==1.22.2
protobuf==3.20.3
pulsar-client==3.2.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypdf==3.9.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rich==13.4.1
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
streamlit==1.22.0
streamlit-chat==0.0.2.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
vertexai==0.0.1
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
